requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
scipy>=1.5.0
pillow>=8.0.0
opencv-python>=4.5.0
transformers>=4.20.0
tensorboard>=2.8.0
tqdm>=4.60.0
PyYAML>=5.4.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
pandas>=1.2.0
networkx>=2.6.0
pycocotools>=2.0.0
